move validator field
