+++ src/renderer/families/cosmos/UndelegationFlowModal/fields/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { default as ValidatorField } from "./Validator";
