+++ src/renderer/families/cosmos/UndelegationFlowModal/types.js
@@ -1,9 +1,7 @@
 // @flow
 import type { Device } from "~/renderer/reducers/devices";
 import type { Step } from "~/renderer/components/Stepper";
-
 import type { Account, TransactionStatus, Operation } from "@ledgerhq/live-common/lib/types";
-
 import type { Transaction } from "@ledgerhq/live-common/lib/families/cosmos/types";
 
 export type StepId = "amount" | "device" | "confirmation";
